+++ changelog.md
@@ -9,7 +9,7 @@
 * Seperated avatar options to /includes/avatar.pug
 * Switched navbar to dark theme
 * Updated title on user profile page to display username
-* Users can now change thier avatar
+* Users can now change thier avatar and username
 
 ## 4/27/19
 * Added SVG avatars